Add pet management fields including HeaderSection, PetNameField, WeightField, BirthDateField, and CategorySection; refactor HomeScreen for database setup
