+++ src/screens/home/HomeScreen.js
@@ -21,10 +21,10 @@ export default function HomeScreen({ navigation }) {
     const setupDatabase = async () => {
       try {
         await initializeDatabase();
-        //await insertMockData();
+        await insertMockData();
         dispatch(fetchPets());
       } catch (error) {
-        //feeding//console.error("Error setting up database:", error);
+        console.error("Error setting up database:", error);
       }
     };
